moved timeAgo function to the backend and created a Comment post-save hook to set timeElapsed property with a relative time string generated from createdAt ISO Date object.
