+++ backend/lib/utils.js
@@ -1,3 +1,5 @@
+const moment = require("moment");
+
 const generateUUID = ({ range }) => {
   const chars =
     "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
@@ -18,10 +20,49 @@ const getCookieToken = (req) => req.cookies.token;
 const emailRegEx = /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,3})+$/;
 const isValidEmail = (email) => emailRegEx.test(email);
 
+const timeAgo = (ISODate) => {
+  moment.updateLocale("en", {
+    relativeTime: {
+      future: "in %s",
+      past: "%s ago",
+      s: (number) => number + " second ago",
+      ss: "%d seconds ago",
+      m: "1 minute ago",
+      mm: "%d minutes ago",
+      h: "1 hour ago",
+      hh: "%d hours ago",
+      d: "1 day ago",
+      dd: "%d days ago",
+      M: "a month ago",
+      MM: "%d months ago",
+      y: "a year ago",
+      yy: "%d years ago",
+    },
+  });
+
+  const secondsElapsed = moment().diff(ISODate, "seconds");
+  const dayStart = moment().startOf("day").seconds(secondsElapsed);
+
+  if (secondsElapsed > 300) {
+    return moment(ISODate).fromNow(true);
+  } else if (secondsElapsed < 60) {
+    return dayStart.format("s ") + "seconds ago";
+  } else {
+    const minute = dayStart.format("m ");
+    const minuteString = minute > 1 ? " minutes ago" : " minute ago";
+    return minute.concat(minuteString);
+  }
+};
+
+const translateISOtoRelativeTime = (ISO) => {
+  return timeAgo(ISO);
+};
+
 module.exports = {
   bool,
   dateToEpoch,
   generateUUID,
   getCookieToken,
   isValidEmail,
+  translateISOtoRelativeTime,
 };
